Update build system, engine core, and configuration
- Enhance CMakeLists.txt with better CUDA and Vulkan SDK detection
- Add shader compilation rules for SPIR-V generation
- Improve Input system with better event handling
- Add Platform.hpp for cross-platform definitions
- Enhance Logger with additional log levels
- Improve save_system with better serialization
- Update settings_manager with validation
- Add AISystem behavior tree improvements
- Enhance ECS with better component management
- Improve JobQueue with thread safety
- Add Math utilities
- Update memory allocators with better tracking
- Improve SceneSerializer with robust loading
- Add .clangd configuration for IDE support
- Add runtime config.json

+++ engine/core/Input.hpp
@@ -19,7 +19,7 @@ public:
     /**
      * @brief Key codes matching GLFW key definitions
      */
-    enum class Key : u16 {
+    enum class Key : i32 {
         Unknown = GLFW_KEY_UNKNOWN,
 
         // Printable keys
@@ -336,7 +336,7 @@ private:
     GLFWwindow* m_window;
 
     // Keyboard state
-    static constexpr u32 KEY_COUNT = static_cast<u32>(Key::KeyCount);
+    static constexpr size_t KEY_COUNT = static_cast<size_t>(Key::KeyCount);
     std::bitset<KEY_COUNT> m_keysCurrentFrame;
     std::bitset<KEY_COUNT> m_keysPreviousFrame;
